Box trace directory split into multiple files
